requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
playsound
gTTS
SpeechRecognition
PyAudio
